+++ tests/Support/Stubs/Events/RequeuableEvent.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Components\Domain\Tests\Support\Stubs\Events;
+
+use Somnambulist\Components\Domain\Events\AbstractEvent;
+
+class RequeuableEvent extends AbstractEvent
+{
+
+}